Fixed Manager-Mr dashboard

+++ src/main/java/com/kavyapharm/farmatrack/dto/DashboardChartsResponse.java
@@ -4,11 +4,10 @@ import java.util.List;
 import java.util.Map;
 
 public record DashboardChartsResponse(
-        List<Integer> salesByMonth,
-        List<Integer> visitsByMonth,
-        List<Integer> targetsByMonth,
-        List<String> monthLabels,
-        Map<String, Integer> expenseByCategory,
-        Map<String, List<Integer>> productSalesByMonth
-) {
+                List<Double> salesByMonth,
+                List<Long> visitsByMonth,
+                List<Double> targetsByMonth,
+                List<String> monthLabels,
+                Map<String, Double> expenseByCategory,
+                Map<String, List<Double>> productSalesByMonth) {
 }
